+++ customer/src/main/kotlin/com/gc/api/customer/adapter/in/graphql/event/EventQueryResolver.kt
@@ -1,6 +1,6 @@
 package com.gc.api.customer.adapter.`in`.graphql.event
 
-import com.gc.api.customer.adapter.`in`.dto.event.request.response.EventResponse
+import com.gc.api.customer.adapter.`in`.dto.event.response.EventResponse
 import com.gc.api.customer.application.service.facade.event.EventFacade
 import com.gc.api.customer.framework.annotation.RequestInfo
 import org.springframework.graphql.data.method.annotation.Argument
